refactor: remove proto from ledger and move to common and update references :hammer:
